+++ src/view/treeNodes/treeNode.ts
@@ -6,7 +6,6 @@
 import * as vscode from 'vscode';
 import Logger from '../../common/logger';
 import { dispose } from '../../common/utils';
-import { FileChangeNode } from './fileChangeNode';
 
 export interface BaseTreeNode {
 	reveal(element: TreeNode, options?: { select?: boolean; focus?: boolean; expand?: boolean | number }): Thenable<void>;
@@ -66,37 +65,6 @@ export abstract class TreeNode implements vscode.Disposable {
 	}
 
 	updateFromCheckboxChanged(_newState: vscode.TreeItemCheckboxState): void { }
-
-	static processCheckboxUpdates(checkboxUpdates: vscode.TreeCheckboxChangeEvent<TreeNode>) {
-		const checkedNodes: FileChangeNode[] = [];
-		const uncheckedNodes: FileChangeNode[] = [];
-
-		checkboxUpdates.items.forEach(checkboxUpdate => {
-			const node = checkboxUpdate[0];
-			const newState = checkboxUpdate[1];
-
-			if (node instanceof FileChangeNode) {
-				if (newState == vscode.TreeItemCheckboxState.Checked) {
-					checkedNodes.push(node);
-				} else {
-					uncheckedNodes.push(node);
-				}
-			}
-
-			node.updateFromCheckboxChanged(newState);
-		});
-
-		if (checkedNodes.length > 0) {
-			const prModel = checkedNodes[0].pullRequest;
-			const filenames = checkedNodes.map(n => n.fileName);
-			prModel.markFiles(filenames, true, 'viewed');
-		}
-		if (uncheckedNodes.length > 0) {
-			const prModel = uncheckedNodes[0].pullRequest;
-			const filenames = uncheckedNodes.map(n => n.fileName);
-			prModel.markFiles(filenames, true, 'unviewed');
-		}
-	}
 
 	dispose(): void {
 		if (this.childrenDisposables) {
